remove deprecated file
