+++ src/Ws.php
@@ -1,7 +1,7 @@
 <?php
 
 /**
- * League.Uri (http://uri.thephpleague.com).
+ * League.Uri (http://uri.thephpleague.com)
  *
  * @package    League\Uri
  * @subpackage League\Uri\Schemes
@@ -17,6 +17,8 @@
 declare(strict_types=1);
 
 namespace League\Uri;
+
+use function class_alias;
 
 /**
  * Immutable Value object representing a Ws(s) Uri.
@@ -73,3 +75,5 @@ class Ws extends AbstractUri
         return $port;
     }
 }
+
+class_alias(Ws::class, League\Uri\Schemes\Ws::class);
