+++ src/functions_include.php
@@ -1,7 +1,7 @@
 <?php
 
 /**
- * League.Uri (http://uri.thephpleague.com).
+ * League.Uri (http://uri.thephpleague.com)
  *
  * @package    League\Uri
  * @subpackage League\Uri\Schemes
@@ -13,7 +13,7 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
+
 if (!function_exists('League\Uri\create')) {
     require __DIR__.'/functions.php';
-    require __DIR__.'/Schemes/deprecated.php';
 }
